refactor(gateway): extract platform response strategies
Collapse the Slack-specific branches in chat-response-bridge.ts into a
PlatformResponseStrategy interface with DefaultResponseStrategy,
SlackResponseStrategy, and TelegramResponseStrategy implementations.
Behavior is preserved byte-for-byte — existing tests pass unchanged.
